Extract `ConnectionSubservice` from `NorisService` and update related dependencies.
