+++ nest-tax-backend/src/noris/noris.module.ts
@@ -3,10 +3,11 @@ import { Module } from '@nestjs/common'
 import { PrismaModule } from '../prisma/prisma.module'
 import ThrowerErrorGuard from '../utils/guards/errors.guard'
 import { NorisService } from './noris.service'
+import {ConnectionSubservice} from "./subservices/connection.subservice";
 
 @Module({
   imports: [PrismaModule],
-  providers: [NorisService, ThrowerErrorGuard],
+  providers: [NorisService,ConnectionSubservice, ThrowerErrorGuard],
   exports: [NorisService],
   controllers: [],
 })
