+++ nest-tax-backend/src/noris/subservices/connection.subservice.ts
@@ -1,0 +1,69 @@
+import { Injectable, Logger } from '@nestjs/common'
+import { ConfigService } from '@nestjs/config'
+import { connect, ConnectionPool } from 'mssql'
+import { ErrorsEnum } from '../../utils/guards/dtos/error.dto'
+import ThrowerErrorGuard from '../../utils/guards/errors.guard'
+
+@Injectable()
+export class ConnectionSubservice {
+  private readonly logger: Logger = new Logger('NorisService')
+
+  constructor(
+    private readonly configService: ConfigService,
+    private readonly throwerErrorGuard: ThrowerErrorGuard,
+  ) {
+    if (
+      !process.env.MSSQL_HOST ||
+      !process.env.MSSQL_DB ||
+      !process.env.MSSQL_USERNAME ||
+      !process.env.MSSQL_PASSWORD
+    ) {
+      throw this.throwerErrorGuard.InternalServerErrorException(
+        ErrorsEnum.INTERNAL_SERVER_ERROR,
+        'Missing one of pricing api envs: MSSQL_HOST, MSSQL_DB, MSSQL_USERNAME, MSSQL_PASSWORD.',
+      )
+    }
+  }
+
+  async createOptimizedConnection(): Promise<ConnectionPool> {
+    const connection = await connect({
+      server: this.configService.getOrThrow<string>('MSSQL_HOST'),
+      port: 1433,
+      database: this.configService.getOrThrow<string>('MSSQL_DB'),
+      user: this.configService.getOrThrow<string>('MSSQL_USERNAME'),
+      connectionTimeout: 60_000,
+      requestTimeout: 180_000,
+      password: this.configService.getOrThrow<string>('MSSQL_PASSWORD'),
+      options: {
+        encrypt: true,
+        trustServerCertificate: true,
+      },
+    })
+
+    return connection
+  }
+
+  async waitForConnection(
+    connection: ConnectionPool,
+    maxWaitTime: number = 10_000,
+  ): Promise<void> {
+    const startTime = Date.now()
+
+    return new Promise((resolve, reject) => {
+      const checkConnection = () => {
+        if (connection.connected) {
+          resolve()
+        } else if (Date.now() - startTime >= maxWaitTime) {
+          reject(
+            new Error(
+              'Connection timeout: Database connection not established within timeout period',
+            ),
+          )
+        } else {
+          setTimeout(checkConnection, 100)
+        }
+      }
+      checkConnection()
+    })
+  }
+}
